Add openshift support
with just the directory ( in fact, just .openshift/action_hooks/deploy ),
zerobin should be usable out of the box on openshift platform, after following
the instruction in README.openshift.md. In short, deploy will just
setup symlink so apache can run and find the needed file.

The paste dump are stored in a directory that should survive reboot
of the instance, but this was not fully tested.

The rest of the files are just placeholder and could be removed.