Create sandboxed webroot
This creates a www webroot directory that can be used as the root for
webserver configuration. The benefit to this approach is that files in
the main directory are no longer exposed on the website. This includes
the data directory, the PHP source code (beyond the required index.php
entry point) and any other files that get placed in the root of the
project during deployment or while running.